Basic Repository structure
The repository type is used to represent a kgit repository on disk. The type encapsulates initialization, validation and other actions that can be taken on a high level .kgit repository.

As kgit is only supposed to work with a single repository at a time, the repository is implemented as a singleton, in a thread-safe manner.