Command and CommandFactory
When one calles kgit the call structure look like:
kgit <COMMAND> <ARGUMENTS>
where arguments are optional, depending on the given
command. In order to have an OO friendly and concise approach, there is a Command type in place, as a sealed construct, that represents a possible command, containing its arguments, as well as being able to execute the command (so itself).
This is well inside the bounds of Single Responsibility, as the arguments are the data required to execute the command, and only the command can utilize those for execution. This is an only-exists-with dependency.

The CommandFactory is responsible for creating those commands, as this involves parsing the command line arguments, and figuring out which command it is.