Main entrypoint
Implementing a basic main entrypoint for the kgit application, to be started with. Right now it's pretty straight forward and simple, by only checking if there are command line arguments, otherwise presenting the caller with an error message and usage instructions.
Right now the main will only create and execute a command.