Repository in current working directory
Except for the init command, all other commands will always assume that the current working directory is the correct directory to work in. Therefore, a global context was created for keeping a defined repository instance, that is used for injecting into the commands. This instance will always reference the current wd.