Restructuring project for repl and cli
Until now the project had only one single main entry point, accepting
command line arguments, like in production usage `kgit status` etc. This
is ok for usages in real szenarios, but in terms of testing or having an
interactive environment this is semi-ideal.
Therefore, the project was restructured into three modules:
1. **ak:** containing the actual code for kgit
2. **cli:** containing the command line executable
3. **repl:** containing the repl for interactive usage

This way, the repl can be used for rapit testing or as an interactive
environment.