Fixing kgitignore not found issue
When invoking kgit status, an error was thrown if the repository didn't
contain a .kgitignore file. As this file is not mandatory, this is not
intendet behavior.

The issue was in the `KGitIgnore` implementation, not checking if the
file actually exists, but assuming it's existing always. Now in case of
non existence an empty set of ignored paths is returned.